+++ src/main/java/dev/isxander/controlify/controller/gyro/GyroComponent.java
@@ -27,6 +27,11 @@ public class GyroComponent implements ECSComponent, ConfigHolder<GyroComponent.C
         return this.config;
     }
 
+    @Override
+    public ResourceLocation id() {
+        return ID;
+    }
+
     public static class Config implements ConfigClass {
         public boolean calibrated = false;
         public boolean delayedCalibration = false;