+++ src/main/java/dev/isxander/controlify/controller/impl/ConfigImpl.java
@@ -3,10 +3,12 @@ package dev.isxander.controlify.controller.impl;
 import com.google.gson.Gson;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
+import dev.isxander.controlify.controller.GenericControllerConfig;
 import dev.isxander.controlify.controller.serialization.ConfigClass;
 import dev.isxander.controlify.controller.ControllerEntity;
 import dev.isxander.controlify.controller.serialization.CustomSaveLoadConfig;
 import dev.isxander.controlify.controller.serialization.IConfig;
+import net.minecraft.resources.ResourceLocation;
 import org.apache.commons.lang3.SerializationException;
 import org.jetbrains.annotations.Nullable;
 
@@ -75,5 +77,10 @@ public class ConfigImpl<T extends ConfigClass> implements IConfig<T> {
     @Override
     public void resetToDefault() {
         this.instance = this.defaultFactory.get();
+    }
+
+    @Override
+    public ResourceLocation id() {
+        return GenericControllerConfig.ID;
     }
 }