+++ src/main/java/dev/isxander/controlify/controller/steamdeck/SteamDeckComponent.java
@@ -1,0 +1,14 @@
+package dev.isxander.controlify.controller.steamdeck;
+
+import dev.isxander.controlify.controller.ECSComponent;
+import dev.isxander.controlify.utils.CUtil;
+import net.minecraft.resources.ResourceLocation;
+
+public class SteamDeckComponent implements ECSComponent {
+    public static final ResourceLocation ID = CUtil.rl("steam_deck");
+
+    @Override
+    public ResourceLocation id() {
+        return ID;
+    }
+}