+++ src/main/java/dev/isxander/controlify/screenop/compat/vanilla/AbstractContainerScreenProcessor.java
@@ -4,7 +4,7 @@ import dev.isxander.controlify.InputMode;
 import dev.isxander.controlify.api.ControlifyApi;
 import dev.isxander.controlify.bindings.ControlifyBindings;
 import dev.isxander.controlify.controller.ControllerEntity;
-import dev.isxander.controlify.controller.dualsense.HapticEffects;
+import dev.isxander.controlify.controller.haptic.HapticEffects;
 import dev.isxander.controlify.gui.guide.ContainerGuideCtx;
 import dev.isxander.controlify.gui.guide.GuideAction;
 import dev.isxander.controlify.gui.guide.GuideActionRenderer;