+++ changelog.md
@@ -7,6 +7,4 @@ This version has the following targets:
 
 ## Bug Fixes
 
-- Fix startup crash on 1.20.1 build.
-- Fix crash when opening the creative inventory. (IF YOU PREVIOUSLY EXPERIENCED THIS, SENDING LATEST.LOG IN MY DISCORD SERVER WITH THIS BUILD WOULD HELP A LOT!)
-- Fix config not loading when a mapping contained a 'I don't have this input' key.
+- Fix Controlify not detecting Steam Deck OLED models